BZ-1079982 - Process deployed to Business-central with PER_PROCESS_INSTANCE session strategy cannot be started via REST (last set of tests)

+++ kie-remote/kie-services-remote/src/test/java/org/kie/services/remote/StartProcessEveryStrategyTest.java
@@ -6,7 +6,7 @@ import org.kie.services.remote.cdi.DeploymentInfoBean;
 public interface StartProcessEveryStrategyTest {
 
     public final static String TEST_PROCESS_DEF_NAME = "org.test.mock.process";
-    public final static int TEST_PROCESS_INST_ID = 4;
+    public final static long TEST_PROCESS_INST_ID = 4;
     
     public void setRuntimeMgrMgrMock(DeploymentInfoBean mock);
     public void setKieSessionMock(KieSession kieSessionMock);
